server/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
Pillow==10.3.0
setuptools==70.0.0
requests==2.32.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
web3==6.20.1